Added buttons, doors and clone limit
A button prefab has been added that is just a trigger with a bool for
when the player/clone is standing on it. In order to use this bool,
other objects must reference the button. This is demonstrated in the
door prefab, which links to a button and uses it to turn on and off. An
example can be seen in level 5.

Additionally, the player now has a clone limit which defaults to 5, be
sure to adjust this limit on each level. When the player spawns a clone
over the limit the earliest clone will be destroyed.