UI Update + Sound Stuff
I added a new font to the assets to use for the in game UI. Added a
bunch of audio sources to objects for easy drag and drop of sounds,
including: die, background, jump, clone, and electric panels. Panel
sound *should* switch on and off depending on the state of the panel.